+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for Bundler.
 - Added ability to duplicate a config object using SimpleConfig::Config#dup (GH-14). [Thanks @andriusch]
 - Added SimpleConfig::Config#merge and SimpleConfig::Config#merge! (GH-14). [Thanks @andriusch]
+- Deprecated SimpleConfig::Utilities.
 
 ## 1.1